# perf(api-rust): switch from image to opencv