--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -806,6 +806,9 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4.2.2
+      - name: Install OpenCV
+        run: |
+          sudo apt-get install libopencv-dev
       - name: Set up Rust
         uses: actions-rust-lang/setup-rust-toolchain@v1.10.1
       - name: Install dependencies