--- a/api-rust/Cargo.toml
+++ b/api-rust/Cargo.toml
@@ -11,7 +11,7 @@ axum = { version = "=0.7.9", features = ["multipart"] }
 dotenvy = "=0.15.7"
 futures-util = "=0.3.31"
 http = "=1.2.0"
-image = "=0.25.5"
+opencv = { version = "=0.93.5", features = ["clang-runtime"] }
 serde = { version = "=1.0.216", features = ["derive"] }
 tch = { version = "=0.18.1", features = ["download-libtorch"] }
 tokio = { version = "=1.42.0", features = ["full"] }