For 1.2 MB image, before the change: 5606 ms:

```sh
2024-12-15T09:55:25.375172Z DEBUG request{method=POST uri=/classify-image version=HTTP/1.1}: tower_http::trace::on_request: started processing request
2024-12-15T09:55:30.981536Z DEBUG request{method=POST uri=/classify-image version=HTTP/1.1}: tower_http::trace::on_response: finished processing request latency=5606 ms status=200
```

After the change: 157 ms

```sh
2024-12-15T10:08:43.378690Z DEBUG request{method=POST uri=/classify-image version=HTTP/1.1}: tower_http::trace::on_request: started processing request
2024-12-15T10:08:43.536464Z DEBUG request{method=POST uri=/classify-image version=HTTP/1.1}: tower_http::trace::on_response: finished processing request latency=157 ms status=200
```